+++ surpyvor/plots.py
@@ -1,9 +1,4 @@
-from cyvcf2 import VCF
 import matplotlib.pyplot as plt
-from surpyvor import utils
-import numpy as np
-from matplotlib_venn import venn2, venn3
-from upsetplot import plot as upsetplot
 
 
 def bar_chart(vcf, outname="stacked_bar.png"):
@@ -11,6 +6,10 @@ def bar_chart(vcf, outname="stacked_bar.png"):
     Make a stacked bar chart for length of the SV split by validation status
     This ignores zygosity.
     """
+    from cyvcf2 import VCF
+    from surpyvor import utils
+    import numpy as np
+
     len_dict = {"True": [], "False": [], "Missed": []}
     for v in VCF(vcf):
         if not v.INFO.get('SVTYPE') == 'TRA' and abs(v.INFO.get('SVLEN')) >= 50:
@@ -48,15 +47,16 @@ def bar_chart(vcf, outname="stacked_bar.png"):
 
 
 def upset_plot(upsets, outname="UpSetPlot.png"):
+    from upsetplot import plot as upsetplot
     upsetplot(upsets, sort_by='cardinality')
     plt.savefig(outname)
 
 
 def venn_diagram(sets, labels, num_samples=2, outname="venn.png"):
     if num_samples == 2:
-        venn = venn2
+        from matplotlib_venn import venn2 as venn
     else:
-        venn = venn3
+        from matplotlib_venn import venn3 as venn
     venn(sets, set_labels=labels)
     plt.savefig(outname)
     plt.close()
